requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=9.5.0
yfinance>=0.2.0
openmeteo-requests>=1.0.0
requests-cache>=1.0.0
retry-requests>=2.0.0
biopython>=1.81
tqdm>=4.65.0
open3d>=0.17.0
